requirements.txt
uvicorn==0.17.6
fastapi==0.78.0
gunicorn==20.1.0
python-multipart
SQLAlchemy==2.0.16
joblib==1.3.2
scikit_learn==1.3.0
azure-core==1.29.2
azure-cosmos==4.5.0
